refactor: implement modular architecture for lib.rs - resolves issue #20
Split monolithic lib.rs (622→84 lines) into focused modules:
- src/analysis/: column analysis, type inference, pattern detection
- src/parsers/: CSV and JSON parsing logic
- src/stats/: numeric and text statistics

✅ Maintains full API compatibility with zero breaking changes
✅ All tests pass, build verified with major features
✅ Improved maintainability and code organization
