+++ src/analysis/mod.rs
@@ -1,0 +1,7 @@
+pub mod column;
+pub mod inference;
+pub mod patterns;
+
+pub use column::analyze_column;
+pub use inference::infer_type;
+pub use patterns::detect_patterns;
