+++ src/parsers/mod.rs
@@ -1,0 +1,5 @@
+pub mod csv;
+pub mod json;
+
+pub use csv::{analyze_csv, analyze_csv_robust, analyze_csv_with_sampling, try_strict_csv_parsing};
+pub use json::{analyze_json, analyze_json_with_quality};
