+++ src/stats/mod.rs
@@ -1,0 +1,5 @@
+pub mod numeric;
+pub mod text;
+
+pub use numeric::calculate_numeric_stats;
+pub use text::calculate_text_stats;
